refactoring: App component was moved to components folder
