--- src/App.module.scss
@@ -1,2 +1,0 @@
-.app {
-}
